Add Intial View
A hook is placed in the application url.py file mapping all URI requests
that are prefixed with /recipes to the main app. Subsquent url mapping
is defined in the recipes/main/urls.py. At this point there is only one
url mapped which points to a hello world view.