Sorted URL Matching
We now use some regex to catch collection sorting ideas. The base url
does not inlcude a sort. We can then sort alphabetically, ingredient,
and category as a whole. In addition a pivot can be applied for these
sorted collections to produce subsets of the sorted collections.

Each url points to the same view but paramertization occurs so wan sort
accordingly.